Refactors quest's InOutOctreeMeshDumper into its own file
